Updated portfolio refactor. Track rev race details and results page yet to be done (along with the endpoint for the results).
