+++ src/app/_components/FooterNote.tsx
@@ -3,7 +3,7 @@ import { FooterNoteProps } from "@/types/portfolio";
 export default function FooterNote({ className }: FooterNoteProps) {
   return (
     <div className="flex items-center justify-center">
-      <p className={className}>© 2024 Udbhav Balaji</p>
+      <p className={className}>© {new Date().getFullYear()} Udbhav Balaji</p>
     </div>
   );
 }
