+++ src/app/_components/home/Profile.tsx
@@ -1,0 +1,23 @@
+function Profile() {
+  return (
+    <div className="container mx-auto my-20 flex flex-wrap items-center justify-center">
+      <p className="mx-auto mb-2 w-full text-center text-lg font-semibold md:text-2xl">
+        Hi there! I'm
+      </p>
+      <h1 className="mx-auto mb-1 w-full text-center text-8xl font-semibold md:mb-3 md:text-8xl">
+        Udbhav Balaji
+      </h1>
+      <p className="mx-auto mb-3 w-full text-center text-lg font-medium md:text-2xl">
+        Software Engineer & Data Analyst
+      </p>
+      <p className="text-md mb-20 mt-10 w-full max-w-3xl text-center font-bold">
+        Welcome to my portfolio! Here are the projects I have built with my
+        various technical skills. I'm always excited to learn new things and
+        work on exciting projects. Check out my work and feel free to give me
+        feedback!
+      </p>
+    </div>
+  );
+}
+
+export default Profile;
